requirements.txt
rerun-sdk
flask
numpy
opencv-python<4.6